+++ app/doc/open_api/v1/users_doc.rb
@@ -17,6 +17,7 @@ class OpenAPI::V1::UsersDoc < OpenAPI::V1::BaseDoc
     param_group :pagination
     param :email, [String, Array], optional: true, desc: 'Filter users by *email* using strict matching.'
     param :user_id, [Integer, Array], optional: true, desc: 'Filter users by *id* using strict matching.'
+    param :created_after, DateTime, optional: true, desc: 'Filter users to accounts created after the given date.'
     example <<-USERS
       # /open_api/v1/users?page=1&per_page=4
       {